Connections now start anonymous
Previously, the :neighbours key of a worker would contain a dictionary
with as key the id of the neighbour. However, in real life, you don't
know the id of the neighbour until you ask. So now the connections start
anonymous, and the id of the neighbour is only known after a greeting.